+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## next
+
+dd mm yyyy
+
+- Change: Apps with the `app-esi` role no longer automatically have access to the default ESI token. Instead, 
+  the corresponding EVE login must now be added to the app (this is done automatically for all existing app
+  with a database update).
+- Implemented EVE login permission check for apps (ESI endpoint) ([#32][i32]).
+
 ## 1.35.0
 
 4 Sep 2022